Moved server types directory and updated documentation
