+++ server/types/AccessLevel.ts
@@ -3,5 +3,5 @@ export enum AccessLevelEnum {
     NONE = -1,
     STANDARD = 0,
     CLUBS,
-    ADMIN
+    ADMIN,
 }
